+++ src/dk/aau/cs/ds306e18/tournament/rlbot/MatchRunner.java
@@ -80,10 +80,12 @@ public class MatchRunner {
                 Path botPackPython = RLBotInstallation.getPathToPython();
                 if (botPackPython != null && Files.exists(botPackPython)) {
                     python = botPackPython.toString();
+                    System.out.println("Found RLBotGUI python installation: " + python);
                 } else {
                     botPackPython = RLBotInstallation.getOldPathToPython();
                     if (botPackPython != null && Files.exists(botPackPython)) {
                         python = botPackPython.toString();
+                        System.out.println("Found RLBotGUI python installation: " + python);
                     }
                 }
             }