+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+#### Version 1.7.3 - August 2020
+- Now supports RLBotGUIX python installations. - NicEastvillage
+
 #### Version 1.7.2
 - Added support three new match options. - NicEastvillage
     - Enable rendering